Create a minimum-viable Go project
This project has a single passing test (1=1 equality) to show that the
test suite is working as expected.

We're using Gom for project dependencies because it allows
dependencies to be vendored easily into the project and doesn't
require the repo to be checked out on the `$GOPATH`. This conforms to
how some development teams expect to work in terms of where they check
their code out.

We're using `make` as the build tool to make the vendor process easier
for testing the project itself.