Rename `publishingApp` struct and make it public for easier reuse
This will be used for the content-store request/response struct as
well. To make this easier we need to make it a public struct.

Also, making it public won't be enough as there's an internal field
called `PublishingApp` already which will lead to issues when trying
to get at the underlying value, hence the rename to
`PublishingApplication`.