Make request logging match expected json format.
This makes the format of request logs match those generated by other
apps.

In order to achieve this we've had to move away from using logrus which
was somewhat opinionated about some of the details.  This implementation
is based on the error logger in the router[1].  This is currently setup
to log to STDOUT because that was the prio behaviour with the logrus
defaults[2].

[1] https://github.com/alphagov/router/blob/master/logger/logger.go
[2] https://github.com/Sirupsen/logrus/blob/master/logger.go#L47