Use the base_path from the URL, not the request body
We have decided that base_path should be omitted from PUT request bodies as it
is already present in the URL and having it twice gives a potential source of
bugs.

Now that we've got a regex capturing the rest of the path, these don't need to
be PathPrefixes and can just be Paths.