ren ContentStoreHandler => ContentStoreController
renaming to convey intent better and not to confuse with:
http://golang.org/pkg/net/http/#Handler

a Handler in Go implements the Handler interface and defines
a ServeHTTP method. you should be able to pass a Handler to
http.Handle("/", ContentStoreHandler) and allow it to handle
requests at that path.

in our case, ContentStoreHandler is not a Handler, but a struct
that has funcs which are later typed to HandlerFuncs in main.go:
http://git.io/AeDs

since we're using it as a controller with functions that are
more or less actions, ContentStoreController is a suitable name.