Live content must be pushed to draft content store
https://trello.com/c/PbCzGBIB

whenever we publish live content, we should replicate it to
to the draft store as well. this ensures search, tagged content
etc continue to work as expected on the draft stack as well.

for now we're not concerned about the response from draft content
store when storing live content. when there is a need to factor that
in, we'll build support for that.