Handle unexpected url-arbiter responses as errors
Until now, this code assumed that if the response wasn't a Conflict (409) or
UnprocessableEntity (422) the it must be successful and that the content can be
pushed to content-store.

This isn't necessarily true - the service might be unavailable (503) or it might
return a response that indicates the request was malformed (400).

In content-store this situation was handled implicitly because these other
errors would be uncaught exceptions and stop the processing. In Go, that doesn't
apply.

I've also added a better error response body than simply serialising the error
object (which might be empty).