Suppress 502 if draft store is not running
an environment variable `SUPPRESS_DRAFT_STORE_502_ERROR`, intended
to be set in development environment, controls whether we pass-on
or suppress 502 returned from the draft content store.

this will not cause failures communicating with publishing-api if
draft content store is not running during development.

there will be a related puppet change to set this variable to true
only in development.