Update makefile to avoid unnecessary gom installs
Previously the VENDOR_STAMP task depended on the
_vendor/src/$(IMPORT_PATH) task.  Because this was a symlink to the root
dir of the repo, its timestamp would be updated avery time the project
was changed, causing make to rebuild both that and the gom install task.

Combining these into one task means that it'll only get run when the
Gomfile is updated.