Refactor content item requests test.
The variables controlling the configuration of the
mock servers were being set by coincidence when a
specific test was run, and leaking to the other tests.

This now sets up each server and default responses
with a default per test, allowing overriding where
necessary.

We also made the following changes:

* contentItem fixture defined in terms of an object rather than a
  string.
* use ghttp.VerifyJSONRepresenting and ghttp.RespondWithJSONEncoded
  which allow us to define our expectations in terms of the same fixture
  object rather than a string
* removed unneccesary expectations
* moved expectations so that they are only defined in tests where they
  are relevant

This code was written by:
- Elliot Crosby-McCullough <elliot.cm@gmail.com>
- David Heath <david.heath@digital.cabinet-office.gov.uk>