Fix processing of url-arbiter responses.
We were only JSON decoding the url-arbiter responses if the Content-Type
header was set to "application/json".  url-arbiter actually sets the
header to "application/json; charset=utf-8", meaning that we ignored the
response body, and returned blank responses on url-arbiter conflicts
etc.

this updates the code in question to use
mime.ParseMediaType(https://golang.org/pkg/mime/#ParseMediaType) to
ensure that the header is processed in a robust manner.