Use relative path in _vendor symlink
This symlink needs to point back at the base of the repo.  Making this a
relative symlink means that when using the dev VM, it will work
consistently between the host and the guest (editor plugins like
goimports will run on the host, and they have issues if the GOPATH isn't
populated correctly).