Log existing commands as events
We want the event log to contain a complete record of all operations performed
through the publishing api so that we could, in theory, replay all of the
events to re-create the whole GOV.UK publishing system in a valid state.

As we evolve the publishing API we expect these commands to eventually be
deprecated. However, for now we want to record these existing commands in the
event log.

This is an inital naive implementation along with request specs, which we
intend to refactor.