Handle validation failures from content store.
Even if the call to content store fails, we should still log the event.

We eventually expect the calls to content store to be done
asynchronously, and in that situation we wouldn't even know if the
content store request had failed. So this approach fits with the
eventual design we'll have.