Event logger records events in transaction
it can take an optional block of code to execute within the transaction.
If the block raises an exception the transaction will be rolled back.

This will be used to ensure that the local derived represenations of the
system state remain in sync with the event history.

For now we are calling out to remote services over HTTP within the
transaction. This has race conditions which mean that the data in the
remote systems may not be in sync with the event log history. We think
that this is ok in this intermediary stage whilst we migrate to the new
way of doing things.

For now, this feature will be used to rollback events if external
services fail or have errors.