Remove global timecop freeze.
Replace it with a teardown to ensure any timecop traveling is reset
after each test.

Having time frozen globally is not obvious when reading/writing tests.
Instead of freezing it globally, we should freeze as necessary for the
specific tests that require it. Doing so will make it more obvious that
the tests are time-sensitive.