Use RSpec metadata to include request helpers.
Instead of relying on a separate top-level helper file that needs to be
required in each spec.

RSpec allows tagging specs, and groups of specs with metadata int he
form of a hash. rspec-rails uses a `type` value to define the farious
spec types. This `type` value can then be used to control where various
helpers are included, which allows for more fine-grained control.