Make request action method name more generic
The `put_content_item` method is part of the api between the test
helpers and the tests themselves. The name `put_content_item` is a bit
too specific as we'd like to use these test helpers for testing other
types of request, such as publishing or putting publish intent.

Therefore rename this method to `do_request`.