Add Replaceable mixin.
When updating an existing live or draft content item, the entity should be
completely replaced by the new data, including blanking any fields which are
not supplied. This change introduces a create_or_replace method, intended to be called within a transaction, that queries for an existing item by content id and
locale, locks it, and completely replaces it by merging the supplied data with
the column defaults for that class. The version number is also updated.