Retry command in case of unique constraint error.
If a request to publishing API times out, the sender will often retry
immediately. This can lead to a case where a second INSERT is made before the
first is committed, leading to a unique constraint violation.

This change attempts to retry the operation three times, in the expectation
that the original request will have completed and the change committed, so that
the second request will see the new entity and update it rather than inserting.

We explicitly roll back the transaction and replay it from the beginning.