Rename content_item to request_body in tests
Make the variable name more generic as sometimes the requests being
exercised in the tests are not sending a content item in the request
body.

Doing this also required parameterising some of the test macros which
make assertions about the logged events to accept an
`expected_attributes` value, rather than making assumptions that the
logged event would match the request body (normally it does, but it
won't always do).