Refactor creates_a_content_item_representation helper
the checks about whether a base path could/couldn't change are not
relevant in all scenarios where this helper is used.

Split those checks out into separate helpers.

Also changed the way of specifying the expected attributes to use a proc
which is `instance_exec`ed. This allows the use of factories and values
defined by `let` blocks in configuring these expectations.