Use Replaceable mixin in LinkSets.
LinkSets need to behave the same as live and draft content items, in that they
need to fully replace existing entities on update, and retry correctly on
unique key violations.

This has involved some refactoring of the Replaceable mixin and specs, to
parameterise the keys used for the unique query and some of the expectations
in the tests themselves.