Rescue the more specific error class
ActiveRecord actually raises a `ActiveRecord::RecordNotUnique` so
there's no need to rescue the error and then check its original type.

Also refactored to use a retry strategy which avoids repeating the
implementation of the code which is surrounded by error handling (so
we're less likely to make a mistake where they could be different).