Adjust replaceable to allow version to be overridden
The versioning behaviour of the old endpoints and the new v2 endpoints
differs. For the v2 endpoints we want the following semantics:

```
PUT /v2/content/:content_id
```

will always increment the version number of the draft item

```
POST /v2/content/:content_id/publish
```

will copy the draft item to the live item without incrementing the
version number.

Therefore we have an invariant condition that if the version number of a
draft item matches the version number of a live item, then that item is
known to be identical.

After discussing the behaviour of the old endpoints we realised that the
old endpoints do not preserve this invariant as currently implemented.

We think that given that the old endpoints should not be mixed with the
new endpoints, this should be ok.

However we should modify the old endpoints to set a version number of
`nil`, and to raise an error if they are used with an item which does
have a version number. This means that:

* we can rely on the invariant remaining true if the version number is
  set
* it will not be possible to switch back to the old endpoints to modify
  a document which has been edited once with the new endpoints (we don't
  believe this should ever happen anyway).