Publish command versioning behaviour
the versioning behaviour of the publish command is different than what
has been implemented for the old v1 endpoints. Therefore I've inlined
the tests which have to do with versioning so that the expected
behaviour can be made explicit.

This change also implements the pact contexts for pact tests and makes
some changes to the data structure of errors to match the expectations
of the pact tests.