Reduce indirection when executing commands
The code had a slight inner-platform effect where
a CommandProcessor was implementing a router from
the API endpoints to the command objects which
execute the work.

This moves the responsibility for this back up to
the controllers.

We also moved the publish action which was incorrectly
located in the v1 controllers.