Allow the creation of linksets without a content item.
We can only currently speculate about the exact
workflows, i.e. whether links might be sent to the
publishing API before the main content item.

Since we don't have a confirmed need for constraining
the order of the pushes, we've opted to leave the
system open (and wrote less code in the process).