Set the version automatically at the model layer...
The draft content item's version auto-increments
and the live content item's version is copied over
from the draft's.

The reason for doing this is so that the app
doesn't need to worry about setting the version
number itself and if it tries to a helpful error
message will be raised.