Validate versioning against the persisted record...
This system handles asynchronous requests to an
API so we should check that the version is always
validated against the persisted record in the 
database.

If we don't do this, it would mean that the
in-memory object could have a stale version which
could cause our version invariants to be violated.