Extract Versioning out of ContentItems
This work has been created as a result of:
https://gov-uk.atlassian.net/wiki/pages/viewpage.action?pageId=40009945

We're exploring whether it would be better to
architect the system in a more extensible way
rather than our ContentItem models becoming too
monolithic.

The hope is that this will make the system easier
to work with because you won't need to consider
every concern at the same time when working on any
particular feature.