Scope the content item by locale when publishing...
We've introduced the capability to specify what
the locale is in the payload. This defaults to
English if one is not provided.

Previously, it would incorrectly choose an arbitrary
draft content item when publishing and this could
cause a live content item in "en" to point
to a draft content item in a different locale.