Validate that base_path is a valid absolute path…
Previously, the url arbiter client was raising an
invalid path error when we didn’t have a model
validation and this was causing the transaction to
rollback.

Now, we raise an error earlier than this via the
model validations. Replaceable doesn’t handle this
very well when we’re associating the version with
the content item. Therefore, we need to check if
the model is valid before we try to save the
version.

I’d be in favour of not using replaceable and
inlining this whole process. Currently, the
indirection of yielding to blocks makes it hard to
trace the order in which records validate and
save relative to each other.