Avoid setting LinkSet#links to nils
When a payload received by the publishing-api V1 did not contain any links, it
used to set LinkSet#links column to nil. Knowing that the default value for the
column is an empty hash this does not feel correct. Moreover it causes issues
after migrating the client to V2, as the code there does not expect nil values
in LinkSet#links columns. I'll address the latter issue in the following
commits.