Do not merge LinkSets' links if they are nil
As explained in the commit titled 'Avoid setting LinkSet#links to nils' the V1
api was setting nil values for the links field on LinSets. After migrating
clients to V2 API this results in an exception as LinkMerger does not expect
nil values in the links field when merging attributes. This in turn results in
`"links": null` posted to the content store that also does not expect nils
in the links field.

Even though as of the previous commit we're defaulting to an empty hash value
for the links field, there are still some legacy records with nils in the
database. This commit makes code a bit more cautious to avoid posting nil values
in the links field to the content store.