Add ContentItemLink table
ContentItemLink table contains the derived representation of the `links` hash of LinkSet.
This table had to be created in order to allow to easily run queries on LinkSet's links.

Source: the item that links to the other item (a content id)
Link type: the type of link (ie `organisations`)
Target: the item that is being linked to (a content id)