Isolate the replaceable context from the model spec…
This was causing tests to fail because the lets
where being evaluated and bleeding instances of
draft content item into the model tests.

I was also hitting a “race condition” error, which
is fixed by this change by explicitly setting a
different base key for the existing content items
to those that are used in the replaceable specs.