Support optimistic locking on v2 endpoints.
This allows publisher applications to detect conflicts
when (for example) a user has left their edit page
open for a long time and someone else has updated
the item in the meantime.

If the version the app thinks it's updating does
not match the version in the Publishing API, a 409
Conflict is returned.

If the `previous_version` key is not provided, no
checking is performed.  This makes this locking
opt-in on the part of the publisher application.

https://trello.com/c/s3dN4eDP/336-optimistic-locking