Fix a race condition when implicitly saving records…
We were seeing unique constraint violations on
productions (for base path) when saving link sets.
This was caused by the fact the saving the version
implicitly saves the associated link set and this
was happening outside of the retry strategy method
call.

This change fixes this problem by ensuring we make
all changes to the the record inside of the retry
strategy. This means that if we decide to save the
record ourselves (explicitly or implicitly) it
will be saved in the context of the retry
strategy and will not raise unique constraint
violations.