Send the version downstream to the content store...
Calls to the content store are now asynchronous and so
they have the potential to arrive out of order.
Because of this, we're sending the version through so
that we can check if messages are out of order.

Caveat: This doesn't happen with the /v1 endpoints
because they may not have a content_id and therefore
aren't persisted in this database. This means that they
don't have an associated version, which is the thing
we're using to check for out of order messages.