Expose version number in get content endpoint...
Versioning is no longer a concern of the content item models, a side
effect of this change is that the version number was no longer exposed
via the get content endpoint, this version number will be used to ensure
optimistic locking of the content item.
This change exposes the version number.