public_updated_at is not required for all content items
Content items without `public_updated_at` could not be saved, because
Array#fetch was being used in the ContentItemPresenter.

public_updated_at is only validated on models when they are renderable,
and redirects are not renderable. This means that ContentItemPresenter
was crashing when `redirect` content items had no `public_updated_at`
field.