Add `/v2/content` endpoint
This commit adds an endpoint to fetch all content-items for a given
format. It will be used to populate the select-boxes in the tagging
interfaces.

For example:

```
/v2/content?content_format=topic&fields[]=title&fields[]=base_path
```

will return:

```
[
{
base_path: "/browse/benefits/entitlement",
title: "Benefits entitlement"
}
]
```

This endpoint is described
https://gov-uk.atlassian.net/wiki/display/FS/RFC+31%3A+Requirements+for+
tagging+architecture.

`placeholder_{format_name}` is used by publishing apps to signal that
content-store should not register routes for a document. The formats
are the same for the purposes of this API, so we can
request both variants.