Update rendering and publishing app factory attributes...
FactoryGirl content item attributes are not representative of real
backend or frontend app names. This is significant because the values
end up in the output of the pacts and consequently in the pact-broker.
For teams wishing to build their own publishing apps this can be
misleading so use some real world example values.