Fix an issue with strings + Postgres’s JSON column type…
By default, assigning a native Ruby string to an
attribute that is backed by a JSON column type,
silently sets the attribute to nil.

When you first #to_json that string, it assigns
the string on the instance, but raises a database
error when saved.

This workaround lets you store Ruby strings in the
JSON column by first wrapping them in an object.
It then unpacks the object in the getter.