Handle publishes of substituted content items…
When a ‘redirect’, ‘gone’ or ‘unpublishing’
content item is published, it would clash with
the existing content item because the base_paths
must be unique in the live content items table.

This fixes that problem by applying the same approach
of “clearing space” for content items in that table.

It’s recognised that this isn’t the ideal thing to
do and that perhaps those “substitute” types of
content items shouldn’t be content items in the
first place. However, this is causing problems on
production and therefore needs a pragmatic fix.