Move the #substitute? method into the SubstitutionHelper...
This keeps all of the substitution logic together in a single
place and means that the model is relieved of any knowledge of
this complicated feature.

It's beneficial to keep this all together because there's
speculation that this will be removed in the future in
favour of a richer "PathReservation" abstraction.