Refactor LinkSetPresenter
This is a change to make future changes easier.

Currently the `LinkSetPresenter` is only used when presenting the links
for the `GET /v2/links/:content_id` endpoint. This commit makes it so
that it is used in all instances where we send the links hash to the
user:

- When setting the links (PUT /v2/links/:content_id)
- When presenting a document (/v2/content/:content_id)

This will make it easier to change what we present as `links` when we
change the way links are stored in the database (#137).