Add Links table
This table will now contain the derived representation of the `links` column of
the `LinkSets` table. This table had to be created in order to allow to easily
run queries on LinkSet's links.

LinkSetId: the id of the source item (foreign key to the `LinkSets`table).
Target: the content id of the item that is being linked to.  Link LinkType: the
type of link (ie `organisations`).