Don’t inline ‘html’ content types in favour of a different approach…
We have decided that publishing api will send the
entire set of content types downstream to both the
content store and onto the message bus.

This should hopefully make the overall system more
flexible in anticipation of change because we’re
not losing information unnecessarily as we send
data downstream.