Protect alpha_taxons from being overwritten
The "alpha_taxonomy" is a prototype for the new taxonomy on GOV.UK. It
is being managed exclusively via the V2 API. While in the migration, we
need to prevent these tags from being overwritten.

It should be noted that this is a temporary "hack", because this code
will be removed with the `/v1` endpoint once all apps have been
upgraded.

Trello: https://trello.com/c/bsZAwZ35