Use ContentItemPresenter in GetContentCollection.
This puts all the presentation logic in the same place, in particular the
calculation of publication_state.

To avoid n+1 queries when getting the versions for publication_state, we do
separate bulk queries on the Version model for both live and draft items; we
can't use the normal `include` ActiveRecord methods because we don't define a
has_one on the content item models in order to keep the Versions separate.