Allow coming soon items to be overwritten
Coming Soon pages are displayed at base_paths that will soon have the actual
content. Whitehall sends the coming soon with a different content-id to the
item that will be published, so until now, publishing-api has been rejecting
the request that tries to publish the actual document.

An alternative might be to change Whitehall to use the same content-id, but
this would mean we'd have to reconcile the existing data in a fairly manual
way.

Once deployed, I think it would be appropriate to republish all documents in
Whitehall that have been rejected by this issue.