Remove access_limited from draft_content_items…
This was made more complicated by the fact that
we’re using the mocks for both POSTing requests to
the API and also for asserting that the right
things are sent downstream to the content store.

It made sense to break the access_limit_params out
of the content items mocks so that they can be
merged in only when POSTing. However, I’d
recommend we go a step further and avoid using 
these mocks as assertions.