Set User.app_name from email by default.
We will be using app_name to restrict the items an app sees by default on the
get_content endpoint, by filtering on publishing_app. GDS-SSO does not send
this data; email and name are not sufficiently constrained to automatically use
either. As a starting point, use the user part of the email address on first
access, this can be changed manually per-app if necessary.