Fake login during tests.
We need two different types of fake login, one for controller specs and one for
requests specs. This is because controller specs interact with Rack::Test
directly, so we have to stub the Warden request environment explicitly,
whereas the request specs go through a higher-level abstraction and we can
simply set GDS::SSO.test_user.