Optimise accessing content by format
Added indexes for format on the content tables and target_id on the
versions table.

Currently the master branch of policy-publisher is undeployable because
calls to this endpoint time out. This is still taking longer than
expected but this commit resolves that issue.

We use this endpoint in policy-publisher to populate select boxes for people,
organisations, and working groups.

I've added a simple benchmark to time this part of the call, and I
also looked at query times logged out by active record when using the
policy publisher.

Some of the queries for the content index endpoint were taking hundreds or
thousands of milliseconds to complete without the indexes, particularly
when the versions were loaded.

I found that rewriting the live content items query made a minor
improvement to the overall time as well. The previous version of the query could
generate very large NOT IN clauses so the execution time was slower than
the other queries.