Allow empty links to be sent
We were sending an empty links hash for pieces of content which have no links
of any kind. This should make it so that those requests don't get rejected.

Additionally, it seems that if the list of links is empty, then all the
existing links should be removed. How else would you remove them?