Add a validator to enforces the CI uniqueness invariant…
Content items should be unique when scoped to the
combination of their supporting objects:

- location (base_path)
- translation (locale)
- state (draft/live)
- semantic version

When the fourth of these objects is created, the
validator kicks in and tries to find any other
content items which have the same set of
supporting objects.

The validation error will be added to that object.