Handle content store errors without response bodies
When content-store can't find a record, it responds with a status code but no
body: https://github.com/alphagov/content-store/blob/master/app/controllers/application_controller.rb#L3-L9

The `error_details` attribute on a `GdsApi::HTTPErrorResponse` is populated
from the response body, so it is nil if the response is empty or not JSON.

This should mean that we no longer get:

  NoMethodError: undefined method `fetch' for nil:NilClass

Though we may now get errors to say that the response was a 404. In the context
of deleting, we should probably swallow 404s as they suggest it was either
never created or already deleted.