Swallow 404s from content-store when deleting
If the content-store returns a 404, it either never had the item or it doesn't
have it any longer. Either way, there's nothing to be done and reporting it to
Errbit isn't helpful.

One qualm I did have is: if a draft is created and quickly discarded, then
delete call might happen before the draft was created. You'd then be in a
situation where the draft was left in the draft content-store. This is already
a problem though; all this change does is to make it no longer report the error
to Errbit.