Make content store push on discard draft more obvious
This change doesn't functionally change anything, because at the point
when the publishing action is called, the live and draft content items
are exactly the same.

However, this change makes the code slightly clearer because it no longer breaks
the pattern "live items go to the live store, draft items to the draft store".