Add delay to `ContentStoreWorker` enqueuing
We are currently experiencing a fair number of
`ActiveRecord::RecordNotFound` errors from the `ContentStoreWorker`.
These are successfully retried so don't impact functionality as such but
do cause a lot of error 'noise'. This commit enqueues the jobs to be run
in `1.second` as a temporary measure to reduce error noise. This is not
meant as a permanent fix. Once the new object model is implemented and
deployed we can revisit if this is still a problem.