Fix a problem with discard draft not cleaning up properly...
Following the deploy to prod, we found that discard draft
should be unconditionally deleting the draft content item.
We should also make sure that a delete request is sent to
the content store because the base path might have changed
between the draft/live content items and the live content
item might not replace the draft in the content store.