A previously published content item might be withdrawn
When saving a draft allow a previously published content item to be in states "published" or "withdrawn" for the following scenario.

When a new content item is published for an existing content item with the same base path the existing published content item is withdrawn. If we eventually try to bring the withdrawn content item back to life the user facing version sequence is reset to 1 because no "published" versions of the content item can be found.

Reusing the same user facing version range becomes a problem as the ContentItemUniquenessValidator finds preexisting duplicate superseded states.