Fix a race condition in PutContent…
Someone had double clicked on the ‘Save & Publish’
button in travel advice publisher. This sent two
requests simultaneously to the Publishing API
which raced against each other.

This change uses row-level pessimistic locking,
which causes the second request to wait for a
mutex before it can proceed. This has the effect
that the second request is processed after the
first completes and we maintain data integrity.

The tests for this change use the truncation
strategy to clean the database instead of the
default transaction strategy. This is because the
threads use separate connections from the
connection pool which cannot be collectively
wrapped in a single transaction.