Inline 'do_request' method calls...
In order to reduce the amount of indirection in tests, this change
inlines the various request helper methods which ultimately rely on a
central 'do_request' method. This does mean more repetition in a couple
of tests, this is a worthwhile tradeoff compared to following the
indirect code path through the request helpers.