Pre-filter by draft/published state...
This query takes about 2 seconds for 08d48cdd-6b50-43ff-a53b-beab47f4aab0.
After adding this restriction, it takes about 100ms. I'm not sure if this
is the right thing to do as it looks like it would affect the total count,
but all of the tests are passing with this line in.