Fix two workflow problems with redirects...
1) Draft redirects were not being created if there was only
a published content item when the PutContent command was
called. In this case, we should look up the previously
published item and compare its base path to the item that
has been put into the system.

2) Previously published items should be withdrawn rather
than superseded if there is a choice between the two. This
happens when an item changes its base path and the item is
then published. Depending on which code runs first in the
Publish command, the previously published item could either
be superseded or withdrawn. This can cause conflicts later
when the item is subsequently redrafted and published.